pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipemind"
version = "0.1.0"
description = "LLM-driven, MCP-integrated, dynamically composable bioinformatics workflow system"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0.1",
  "pydantic>=2.7",
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "typer>=0.12",
  "rich>=13.7",
  "networkx>=3.3",
  "snakemake>=8.11",
  "fastmcp>=0.4.1",
  "httpx>=0.27",
  "openai>=1.0.0",
]

[project.scripts]
pipemind = "pipemind.cli.app:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
# Ship the generated registry by default so the server works out-of-the-box after install.
pipemind = [
  "registry/*.yaml",
  "registry/*.yml",
  "registry/README.md",
]
